LPS-58744 Move to JornalArticleLocalServiceTest since it's this service we're testing
